+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## [Unreleased]
 
+### Added
+- ROG Arion external driver LED support
+- Add GA605W LED layout
+
 ### Changed
 - Fix attribute writes. At some point the kernel API seems to have changed.
 - Extremely large refactor of Aura device handling. Should enable easy add of different kinds now.
-- Add GA605W LED layout
 - Rename CLI args for aura related properties. This will likely change further as more devices are added
 
 ## [v6.0.12]